New Recents: Fix color swap bug
This is though. First, set bar colors every a view is
created/updated. Before, it was only set when loading
TaskDescription was finished, and never touched on updating
(=> wrong colors).

Second, when changing recent item's ordering,
TaskDescriptions are refreshed, which causes all of them to
be destroyed completely and recreated afterwards. Thus, old
AB color information is not available anymore and we need
reevaluate it. When now setting colors on every view update,
often times the corresponding TaskDescription still hasn't
finished loading and we therefore need to fall back to
default values (very ugly). The solution now is to hand over
old TaskDescriptions on every refresh and to reuse existing
AB color information during recreation of the new
TaskDescriptions.