Detect when dynamic navigation bar shouldn't be on
This commit adds a simple check to see when the navigation bar height
is set to a non-positive value (<= 0) which is then checked against
whether the navigation bar is set to be dynamic. If both are true,
the dynamic navigation bar is automatically reset to "disabled". This
should be handled in the settings app primarely so this is just a
helper.

Change-Id: Id750fab7cd1371660588588ee35642eefdaa0d56