Make DSB make less checks
This commit makes DSB even lighter on the system on average. If no
color updates take place and everything is working as expected, the
checker just backs up to a delay of 2 seconds between checks. While
watching videos, the checks will still remain consistently fast and
will provide a fluid experience, but during more casual usage the
delay of 2 seconds can be accepted. Additionally to the backoff,
this commit also makes the checks be stalled completely while the
screen is turned off. We do not need to update the status bar
during that period at all.

Change-Id: I5657a59b41b0f1557f69ae9aaf34ab52d28678d2