Bump up the minimal DSB update frequency
This commit brings up the minimal DSB update frequency to 1x per
second. This is twice as fast on the slowest speed compared to what
was previously set but having it as once each 2 seconds does feel a
bit sluggish at times. Should do testing to see if this is acceptable
system resource usage.

Change-Id: I8ac2fec26ef13c52cb7be4eaff9b8602d184d589