Introduce modes for DSB
This commit adds a gradient setting and a darkened status bar setting
to DSB. Along with that, all the settings/configuration is now updated
via observers. All that is essentially left is to use resources for
the text coloring to allow themeability of them (see clock or ticker
texts) when no_override is set.

Change-Id: I25cd671570ce93fa9b7ae019263332aa366bfd1e