Switch DSB a level lower using JNI [no release]
This commit moves the same implementation we had but just takes it a
level lower - to native code. A bunch of calculations are still done
in Java and the general implementation is still essentially the same.

This is not release ready because lanscape is broken again (the
image does not appear to be pre-rotated for us) and the red and
blue channels appear to be swapped for me. My initial guess is
that the latter issue could be related to endianness.

Change-Id: I2440afcb76fa17fd72e1f326c515cce0e12e28a3