Make DSB correctly use the red and blue channels
This commit fixes the bug of the red and the blue channels being
swapped which was introduced when screenshot/color grabbing was
switched over to C++. The bug was really related to endianness as
initially assumed but instead was due to the RGBA value being stored
as BGRA behind the scenes. Java code uses big endian so it is ARGB
instead. If someone has their red and blue channels swapped, their
device must be on big endian by default or something and this will
need revisiting then, but for now, it should be fixed.

Change-Id: I4aba09ce7429ea5d12deebb6efd302e66eae404f