Tweak system bar background invalidation
This commit tweaks the method used for invalidating the system bar
background drawables and ensures they are not being forced to be
redrawn mutliple times at the same point in time. The animation
starting and ending times receive minor changes in this commit as
well to follow the invalidation logic changes. This deals with some
of the lag experienced on the system bars but we still need changes
to kick down the CPU requirement.

Change-Id: Ia1e7f718f93156f5aed547ede3a1718ca8f3a433