Work on maintaining the navigation bar DSB state
This commit makes the navigation bar view hold on to the override
button color passed in by the bar background updater. This could
fix the reports of the navigation bar buttons being the incorrect
state but as I have no way to consistently reproduce the bug I
can't be sure.

Change-Id: I8ebd3784a36930ed90a2ca3c07b1f2580d9fdd87