Fix DSB panel detection and landscape
This commit kind of rewrites how color are picked from the grabbed
screenshot. As a result of this rewrite, landscape is functioning
much more as one would expect (without a mind of its own) and panel
detection should be improved by this on both sides of the screen as
well. The right side of the screen is picked as the dominant one
because panels that slide open usually come in from the left side
and we want to show the background color on system bars unless we
also make the opened panel extend across the system bars fully.

Change-Id: Id525ff1d74bf2138adb7f49aecb98764fafffef1