Make the DSB screenshots be full screen
This commit switches the screenshot grabbing logic around to be full
screen instead of using 200 cropped pixels for height. This makes the
grabbing take more time and use up more system resources as this is
pretty much a revert of the area grabbing code but this makes the
colors get picked properly. We care about that a lot as well as having
speed.

Change-Id: I8a9a3960d3ddbb6c709a770dd140063d6e574093