Patch up the DSB optimization
This commit is centered around cleaning up the massive optimization
commit. Changes happen all over the place really. Some of the
highlights are: making the listeners receive the previous override
color in turn allowing simple animation logic which does not require
messing around with tags or using other hacks; removing the 800 by 800
testing screenshot size as it goes over the size of some devices which
in turn means it is not really safe for them.

Change-Id: I9c9ad8599df2a410e47f584d15e30267e2bbd396