Recents View: fix mIsScrolling not being set correctly
Actually, the whole touch handling is quite a mess and this fix is not really good, but the best I could come up with.
The problem is, that in onDispatchTouchEvent onTouchEvent is being called therefore onTouchEvent is being called twice, if onInterceptTouchEvent returns true.
If onInterceptTouchEvent isn't called at all, because no child is touched, mIsScrolling isn't set but onTouchEvent is because of dispatchTouchEvent.

Change-Id: I881af1041eafe33e5d3ce2bf09547ac36cd35bba