Recents View: fix swipe to the side doesn't work since the last fix
Because mIsScrlling wasn't set to false on ACTION_UP, swiping to stop a process didn't work anymore after scrolling.
Setting it to false in ACTION_UP on the other hand caused thumbnails to be clicked even if they weren't supposed to be clicked.
This is now fixed using the new mWasScrolling variable.

Change-Id: I4f66688ec7cce668c872c9c0e21482be9f4fe2a0